changed module

+++ pokemon-teams-backend/app/controllers/pokemons_controller.rb
@@ -11,14 +11,15 @@ class PokemonsController < ApplicationController
   end
 
   def create
-    pokemon = Pokemon.new(allowed_params)
-
-    if pokemon.valid?
+    # we can createa a validator here by looking for the trainer params[pokemon][trainer_id] and check the size of its team
+    pokemon = Pokemon.new(nickname: Faker::Name.first_name, species: Faker::Games::Pokemon.name, trainer_id: params[:trainer_id])
+    
+    if pokemon.valid? && pokemon.trainer.valid?
       pokemon.save
-      
+      render json: pokemon, status: 200
     else
       flash[:error] = pokemon.errors.full_messages
-      render json: flash[:error]
+      render json: flash[:error], status: 404
     end
 
   end
@@ -26,13 +27,14 @@ class PokemonsController < ApplicationController
 
   def destroy
     pokemon = Pokemon.find(params[:id])
-    pokemon.destroy 
+    unless pokemon.nil?
+      pokemon.destroy
+      render json: pokemon
+    else
+      render json: { error: "Pokemon not Found!" }, status: 404
+    end
     
   end
 
-  private 
 
-  def allowed_params
-    params.require(:pokemon).permit(:nickname, :species, :trainer_id)
-  end
 end
